Fix issue with cropper sometimes not opening (#646)
This fixes an issue where sometimes the cropper won't open when clicking the Kap tray icon until you open the About or Preferences window. 

It was originated here: https://github.com/wulkano/kap/pull/630/files#diff-8a409ab0e00c53b13c315c26ebe40be8R82

I was trying to get it so it wouldn't close/reopen the cropper when you click the accelerator multiple times, but for some reason if the cropper windows don't close normally, it gets in a bad state where that boolean remains true even though the croppers are not open. So, when you try to open them it returns and does nothing. I moved the check to the accelerator itself and used the actual cropper Map instead of a boolean to check if the croppers are open.